Add working function to set updated volume for user optimistically
